🐛 fix: an integer literal takes the width of what it is written into
Inference types every integer literal as the default `int`, and nothing
afterwards reconsidered it. A literal only ever reached a wider slot by
being bound to an annotated `let` first, so `List<i128>([3, 1, 2])` built
its elements as 8-byte ints and copied them into 16-byte slots, and
`l.set(0, <past i64::MAX>)` was refused outright with a bound the source
never asked for.

The float widths already solve exactly this, because only the consuming
context knows the width: `narrow_float_literals` walks the literal, a
sign in front of it, a collection literal's elements and a constructor
wrapping one, and records the width at each. There was no integer twin.
There is now, wired at the same sites, plus the collection constructor
whose element type comes from its type argument rather than a declared
parameter.

The range check could not simply consult the target, because at an
argument position the literal is inferred before the parameter type is
resolved — the error had already been reported by the time the width was
known. It is now held and judged once the bodies are checked, against the
type finally recorded for the literal, which is also what lets the
message name that type instead of the default.

Two things had to follow, both found by the suite rather than by
reasoning. A negated literal now reached a 128-bit negation, which
Cranelift cannot lower — so a literal's sign is applied while compiling,
where it was already known, and no negation instruction is emitted.
Small literals now legitimately reach 128-bit slots, where `iconst` does
not exist — so a 128-bit constant is built from its halves whatever the
literal's own width, rather than only for the 128-bit variants.

This closes the sign flip filed separately as the negated-literal defect:
the magnitude no longer travels through a 64-bit constant, so
`let n i128 = -9223372036854775809` keeps its sign.

Snapshots that recorded the old shape are updated: a negated literal is a
constant rather than a `Neg`, and an annotated declaration needs no cast
to reach its declared width. `return 10` in an `i64` function still
casts — the return position has no narrowing site, for floats either, and
that gap is filed rather than widened into here.

One test changed meaning. `test_enum_i128_payload_with_second_field`
asserted that a 128-bit enum payload does not overrun the fields after
it, and passed only because the literal it used was silently an `int`, so
no 128-bit payload was ever stored. It stores one now, and the second
field reads back the first one's value. A compiler built at ec6d85cc
fails the same program when the payload is a declared binding, so the
defect predates this change; the test is left ignored, naming what it
found, for the task that owns it.

+++ tests/mir/lowering/constant.rs
@@ -31,7 +31,7 @@ fn const_typed_integer() {
 
             bb0: {
                 StorageLive(_1);
-                _1 = const Integer(I8(42)) as i32;
+                _1 = const Integer(I8(42));
                 StorageDead(_1);
                 return;
             }
